+++ realnvp/__init__.py
@@ -1,5 +1,5 @@
 """realnvp - Basic RealNVP implementation in jax."""
 
-__version__ = '0.1.0'
-__author__ = 'Daniel Ward <danielward27@outlook.com>'
+__version__ = "0.1.0"
+__author__ = "Daniel Ward <danielward27@outlook.com>"
 __all__ = []
