+++ realnvp/bijections/coupling.py
@@ -1,0 +1,57 @@
+from realnvp.bijections.abc import ParameterisedBijection
+import equinox as eqx
+import jax.numpy as jnp
+from jax import random
+
+
+class Coupling(eqx.Module):
+    d: int  # Where to partition
+    D: int  # Total dimension
+    bijection: ParameterisedBijection
+    conditioner: eqx.nn.MLP
+
+    def __init__(
+        self,
+        key: random.PRNGKey,
+        bijection,
+        d: int,
+        D: int,
+        conditioner_width: int,
+        conditioner_depth: int,
+    ):
+        self.d = d
+        self.D = D
+        self.bijection = bijection
+        output_size = self.bijection.num_params(D - d)
+        self.conditioner = eqx.nn.MLP(
+            d, output_size, conditioner_width, conditioner_depth, key=key
+        )
+
+    def __call__(self, x: jnp.ndarray):
+        return self.transform_and_log_abs_det_jacobian(x)
+
+    def transform_and_log_abs_det_jacobian(self, x):
+        x_cond, x_trans = x[: self.d], x[self.d :]
+        bijection_params = self.conditioner(x_cond)
+        bijection_args = self.bijection.get_args(bijection_params)
+        y_trans, log_abs_det = self.bijection.transform_and_log_abs_det_jacobian(
+            x_trans, *bijection_args
+        )
+        y = jnp.concatenate([x_cond, y_trans])
+        return y, log_abs_det
+
+    def transform(self, x):
+        x_cond, x_trans = x[: self.d], x[self.d :]
+        bijection_params = self.conditioner(x_cond)
+        bijection_args = self.bijection.get_args(bijection_params)
+        y_trans = self.bijection.transform(x_trans, *bijection_args)
+        y = jnp.concatenate([x_cond, y_trans])
+        return y
+
+    def inverse(self, y: jnp.ndarray):
+        x_cond, y_trans = y[: self.d], y[self.d :]
+        bijection_params = self.conditioner(x_cond)
+        bijection_args = self.bijection.get_args(bijection_params)
+        x_trans = self.bijection.inverse(y_trans, *bijection_args)
+        x = jnp.concatenate([x_cond, x_trans])
+        return x
