+++ realnvp/bijections/rational_quadratic_spline.py
@@ -1,0 +1,94 @@
+from realnvp.bijections.abc import ParameterisedBijection
+import jax.numpy as jnp
+import jax
+
+
+class RationalQuadraticSpline1D(ParameterisedBijection):
+    K: int
+    B: float
+    _pos_pad: jnp.ndarray  # End knots and beyond
+
+    def __init__(self, K, B):
+        self.K = K
+        self.B = B
+        pos_pad = jnp.zeros(self.K + 4)
+        pad_idxs = jnp.array([0, 1, -2, -1])
+        pad_vals = jnp.array(
+            [-B * 1e4, -B, B, B * 1e4]
+        )  # Avoids jax control flow for identity tails
+        pos_pad = pos_pad.at[pad_idxs].set(pad_vals)
+        self._pos_pad = pos_pad
+
+    @property  # TODO update to Buffer when introduced to eqx: https://github.com/patrick-kidger/equinox/issues/31
+    def pos_pad(self):
+        return jax.lax.stop_gradient(self._pos_pad)
+
+    def transform(self, x, x_pos, y_pos, derivatives):
+        # Roughly following notation from Neural Spline Flow paper (univariate)
+        k = jnp.where((x <= x_pos[1:]) & (x > x_pos[:-1]), size=1)[0][0]
+        yk, yk1, xk, xk1 = y_pos[k], y_pos[k + 1], x_pos[k], x_pos[k + 1]
+        sk = (yk1 - yk) / (xk1 - xk)
+        xi = (x - xk) / (xk1 - xk)
+        num = (yk1 - yk) * (sk * xi ** 2 + derivatives[k] * xi * (1 - xi))
+        den = sk + (derivatives[k + 1] + derivatives[k] - 2 * sk) * xi * (1 - xi)
+        return yk + num / den
+
+    def inverse(self, y, x_pos, y_pos, derivatives):
+        k = jnp.where((y <= y_pos[1:]) & (y > y_pos[:-1]), size=1)[0][0]
+        xk, yk = x_pos[k], y_pos[k]
+        sk = (y_pos[k + 1] - yk) / (x_pos[k + 1] - xk)
+        y_delta_s_term = (y - yk) * (derivatives[k + 1] + derivatives[k] - 2 * sk)
+        a = (y_pos[k + 1] - yk) * (sk - derivatives[k]) + y_delta_s_term
+        b = (y_pos[k + 1] - yk) * derivatives[k] - y_delta_s_term
+        c = -sk * (y - yk)
+        xi = (2 * c) / (
+            -b - jnp.sqrt(b ** 2 - 4 * a * c)
+        )  # TODO how to deterimine correct root?
+        x = xi * (x_pos[k + 1] - xk) + xk
+        return x
+
+    def transform_and_log_abs_det_jacobian(self, x, x_pos, y_pos, derivatives):
+        raise NotImplementedError
+
+    def num_params(self, dim: int = None):
+        return self.K * 3 - 1
+
+    def get_args(self, params):
+        widths = jax.nn.softmax(params[: self.K]) * 2 * self.B
+        heights = jax.nn.softmax(params[self.K : self.K * 2]) * 2 * self.B
+        derivatives = jax.nn.softmax(params[self.K * 2 :])
+        derivatives = jnp.pad(derivatives, 2, constant_values=1)
+        x_pos = jnp.cumsum(widths) - self.B
+        x_pos = self.pos_pad.at[2:-2].set(x_pos)
+        y_pos = jnp.cumsum(heights) - self.B
+        y_pos = self.pos_pad.at[2:-2].set(y_pos)
+        return x_pos, y_pos, derivatives
+
+
+class RationalQuadraticSpline(ParameterisedBijection):
+    K: int  # Number of inner knots
+    B: int  # Bound size
+    spline: RationalQuadraticSpline1D
+
+    def __init__(self, K, B):
+        """As for 1D spline, but we vmap over x dimensions. i.e. parameter args
+        are now matrices, with each row being parameters for each dimension of x."""
+        self.K = K
+        self.B = B
+        self.spline = RationalQuadraticSpline1D(K, B)
+
+    def transform(self, x, x_pos, y_pos, derivatives):
+        return jax.vmap(self.spline.transform)(x, x_pos, y_pos, derivatives)
+
+    def transform_and_log_abs_det_jacobian(self, x, x_pos, y_pos, derivatives):
+        return jax.vmap(self.spline.transform)(x, x_pos, y_pos, derivatives)
+
+    def inverse(self, y, x_pos, y_pos, derivatives):
+        return jax.vmap(self.spline.inverse)(y, x_pos, y_pos, derivatives)
+
+    def num_params(self, dim: int):
+        return (self.K * 3 - 1) * dim
+
+    def get_args(self, params):
+        params = params.reshape((-1, self.K * 3 - 1))
+        return jax.vmap(self.spline.get_args)(params)
