--- realnvp/nn.py
@@ -1,2 +1,0 @@
-
-class MaskedMLP(eqx.Module)
